+++ demo/Demo.tsx
@@ -10,7 +10,7 @@ import {
   Tabs, Tags, DropZone, Textarea, DateTimePicker, TimePicker,
   DateTimeRange, TimeRange, DropDownLink, Currency, Rating,
   Div, Table, Dl, FileUpload, HTMLTags, Validation,
-  Collapsible, Collapse, VStepper, Button, Tooltip, FileDrop, Tour,
+  Collapsible, Collapse, VStepper, Button, Tooltip, FileDrop, Tour, Wizard
 } from './components';
 import { useElementRef } from '../leda/utils';
 
@@ -73,6 +73,7 @@ export const Demo = hot(() => (
       <Tour />
       <Validation />
       <VStepper />
+      <Wizard />
     </L.Div>
   </L.Div>
 ));