+++ leda/components/Modal/handlers.ts
@@ -2,7 +2,13 @@ import * as React from 'react';
 import { ModalWindowProps } from './types';
 
 export const createEscapePressHandler = (props: ModalWindowProps) => (ev: React.KeyboardEvent<HTMLElement>): void => {
-  const { onEscapePress, onClose } = props;
+  const {
+    activeAlertKey,
+    onEscapePress,
+    onClose,
+  } = props;
+
+  if (activeAlertKey != null) return;
 
   const isEscapeKey = ev.key === 'Escape' || ev.key === 'Esc';
 
@@ -21,7 +27,13 @@ export const createCloseButtonClickHandler = (props: ModalWindowProps) => (ev: R
 };
 
 export const createOverlayClickHandler = (props: ModalWindowProps) => (ev: React.MouseEvent<HTMLElement>) => {
-  const { onOverlayClick, onClose } = props;
+  const {
+    activeAlertKey,
+    onOverlayClick,
+    onClose,
+  } = props;
+
+  if (activeAlertKey != null) return;
 
   const elementThatCursorIsOver = document.elementFromPoint(ev.clientX, ev.clientY);