+++ leda/components/Modal/index.ts
@@ -1,4 +1,5 @@
 export { Modal } from './Modal';
-export { ModalHeader } from './ModalHeader';
+export { ModalAlert } from './ModalAlert';
 export { ModalBody } from './ModalBody';
 export { ModalFooter } from './ModalFooter';
+export { ModalHeader } from './ModalHeader';