+++ leda/index.ts
@@ -60,7 +60,11 @@ import { Main } from './components/Main';
 import { Mark } from './components/Mark';
 import { MaskedInput } from './components/MaskedInput';
 import {
-  Modal, ModalHeader, ModalBody, ModalFooter,
+  Modal,
+  ModalAlert,
+  ModalBody,
+  ModalFooter,
+  ModalHeader,
 } from './components/Modal';
 import { MultiSelect } from './components/MultiSelect';
 import { Nav } from './components/Nav';
@@ -96,6 +100,7 @@ import { Ul } from './components/Ul';
 import { Leda, LedaContext, UnderscoreClasses } from './components/LedaProvider';
 import { validate } from './validators';
 import { VStepper } from './components/VStepper';
+import { Wizard } from './components/Wizard';
 
 import * as Validation from './components/Validation';
 import * as AutoCompleteTypes from './components/AutoComplete/types';
@@ -232,7 +237,11 @@ export {
   Main,
   Mark,
   MaskedInput,
-  Modal, ModalHeader, ModalBody, ModalFooter,
+  Modal,
+  ModalAlert,
+  ModalBody,
+  ModalFooter,
+  ModalHeader,
   MultiSelect,
   Nav,
   Notifications,
@@ -265,6 +274,7 @@ export {
   Leda, LedaContext, UnderscoreClasses,
   validate,
   VStepper,
+  Wizard,
   form,
   utils,
   Validation,